refactor: saving/application/ 구조 개선
- 기능 중심으로 패키지 분류
- 파일 리네임
