+++ main-server/src/main/java/com/freedom/saving/api/controller/SavingSubscriptionCommandController.java
@@ -5,11 +5,11 @@ import com.freedom.common.logging.Loggable;
 import com.freedom.common.security.CustomUserPrincipal;
 import com.freedom.saving.api.dto.OpenSubscriptionRequest;
 import com.freedom.saving.api.dto.OpenSubscriptionResponse;
-import com.freedom.saving.application.service.SavingSubscriptionCommandService;
-import com.freedom.saving.application.OpenSubscriptionCommand;
-import com.freedom.saving.application.OpenSubscriptionResult;
-import com.freedom.saving.application.SavingSubscriptionService;
-import com.freedom.saving.application.service.MaturitySettlementService;
+import com.freedom.saving.application.subscription.SavingCancelService;
+import com.freedom.saving.application.subscription.dto.OpenSubscriptionCommand;
+import com.freedom.saving.application.subscription.dto.OpenSubscriptionResult;
+import com.freedom.saving.application.subscription.SavingOpenService;
+import com.freedom.saving.application.maturity.MaturitySettlementService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,8 +25,8 @@ import java.math.BigDecimal;
 @RequiredArgsConstructor
 public class SavingSubscriptionCommandController {
 
-    private final SavingSubscriptionService service;
-    private final SavingSubscriptionCommandService commandService;
+    private final SavingOpenService service;
+    private final SavingCancelService commandService;
     private final MaturitySettlementService maturitySettlementService;
 
     public record MaturityQuoteResponse(BigDecimal principal, BigDecimal rate, BigDecimal interest, BigDecimal total) {}
