+++ main-server/src/main/java/com/freedom/saving/api/dto/MaturityPreviewResponse.java
@@ -1,6 +1,6 @@
 package com.freedom.saving.api.dto;
 
-import com.freedom.saving.application.service.SavingMaturityCalculationService.MaturityCalculationResult;
+import com.freedom.saving.application.maturity.SavingMaturityCalculationService.MaturityCalculationResult;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
