--- main-server/src/main/java/com/freedom/saving/application/dto/SavingProductOptionItem.java
@@ -1,21 +1,0 @@
-package com.freedom.saving.application.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-
-/**
- * 상세 옵션(기간/금리) dto
- */
-@Getter
-@Setter
-public class SavingProductOptionItem {
-
-    private Integer termMonths;
-    private BigDecimal rate;
-    private BigDecimal ratePreferential;
-    private String rateType;
-    private String rateTypeName;
-
-}
