+++ main-server/src/main/java/com/freedom/saving/application/query/dto/SavingProductListItem.java
@@ -1,4 +1,4 @@
-package com.freedom.saving.application.dto;
+package com.freedom.saving.application.query.dto;
 
 import lombok.Getter;
 import lombok.Setter;
